Fixes #44. Handles errors in output of du command
Currently errors from the du command are ignored
which can cause a panic.

This commit adds the x flag for du which  skips
directories on different file systems. This
gives a more correct value for disk usage as
well as avoiding errors such as:
`No such file or directory` for a file in a
mounted volume.